Subliminal now runs on the device.
By having UIAutomation sends responses using UIAApplication.setPreferencesValueForKey(),
instead of manually locating and writing to the NSUserDefaults plist (which could work in the simulator).

Unfortunately, UIAApplication.setPreferencesValueForKey() does not use the NSUserDefaults plist
in the simulator, but rather the host domain-level plist (see
http://stackoverflow.com/questions/4977673/reading-preferences-set-by-uiautomations-uiaapplication-setpreferencesvaluefork),
so now SLRadio.js is much simpler
but we've got to do a bit of work to locate the appropriate plist in SLTerminal.m.