The application under test may now register arbitrary actions with the SLTestController, to be invoked by tests.
This allows the tests to manipulate application state directly
(as opposed to through UIAutomation), while avoiding coupling
to the application structure.

This mechanism does not retain the targets, and actions are performed
on the main thread.

Actions may return values. If these values are objects, they themselves are proxified
so that they too may be safely used off the main thread.

An example has been added to STLoginTest.