Refactored accessibility methods on NSObject
Moved the accessibility reflection methods into NSObject to have parity with the definitions in UIKit/UIAccessibility. This removes assumptions about what is a UIAccessibilityElement and what has a UIAccessibilityContainer. Element lookup now returns the object chain needed to access the element so the chain does not have to be rebuilt afterwards.