Replaced the subliminal button with preferences keys
The preferences file is now used to communication in both directions. The app and script execute in lock-step order by waiting for each other to update their respective keys. The scripts polls the "commandIndex" key and waits for it to increment before eval'ing the "command" key. The app waits for the result by polling for the existence of the "resultIndex" key. The app then checks the "result" and "exception" keys for the result of eval().