Add currentUIAPrefix debugging method to SLElement
Adds a new method to SLElement, currentUIAPrefix, which does the
work of getting the uiaPrefix of the element under the hierarchy
of the current main window.  Unlike uiaPrefix, which dispatches to
the main thread and retries a few times, currentUIAPrefix simply
executes synchronously and returns a result, making it suitable for
use with the debugger.