Skip elements in the accessibility chain that refuse accessibilityLabel/Identifier
There are elements that appear in the accessibility hierarchy that
do not respect our code's setting of accessibility properties.  When
we encounter such an object in currentUIAPrefix we can't give it
any sort of identifier that the automation instrument will recognize.

We have found, empirically, that simply skipping these troublesome
objects usually results in a chain that the automation instrument
will correctly interpret, so that's what we're going to do.  Just
skip any elements for which we cannot get a non-nil accessibility
name.