Does not include all elements in uiaPrefix fixes #36
Alters the creation of uiaPrefix to not include elements without
accessibility names. Does not create accessibility identifiers for
elements, except in the case of UIScrollViews and UIImageViews.
An exception is made for these classes because unlike other classes,
they are added to uiautomations accessibility hierarchy even if they do not
have an accessibility name. Assigning them an identifier ensures
they will be included in the prefix created by Subliminal, and that
UIAutomation will be able to successfully traverse the accessibility
hierarchy to identify the element to operate on.