Should not cast objc_msgSend to return NSCopying for void return type
We were incorrectly casting objc_msgSend to a function of type
id<NSCopying>(*)(id, SEL, id) when the selector argument had a
void return type.  We should be using "void(*)" for the return
type.  Doing otherwise leads to crashes because the compiler will
insert a call to attempt to retain the (void) return value.