Include an explicit drag duration much faster than the UIAutomation default
Adds a duration option to the dragInsideWithOptions method that we
generate for dragging in SLElements.  The default duration for drag
actions, used if we do not include a duration option, is 1 second,
and in practice the drags actually seem to take a little longer
than that.  If we set the duration to 0.2 seconds then the drags
happen at a much more reasonable speed and are comparable to other
animations used throughout iOS.