Use accessibility identifier replacement on all UIControls
Gives all UIControls that do no already have an accessibility label
or accessibility identifier a unique (based on the class and address
of the object) accessibilty identifier.  When recording events,
UIAutomation includes controls in accessibility chains even when
the controls do not have accessibility labels or identifiers, so
we need to make sure that Subliminal also includes those controls.

Previously we used this replacement behavior on UISegmentedControl,
but UISegmentedControl inherits from UIControl so we no longer need
an implementation for UISegmentedControl in particular.