Add new SLElement subclasses for keyboard and keyboard keys (resolves #48)
Adds new subclasses of SLElement that return strings from uiaSelf
that will match the keyboard or sub elements of the keyboard.
Both new classes have methods that initialize an SLElement with
a nil matching predicate.  This is OK because the uiaSelf
overrides prevent matchesObject from ever being called on instances
of the new classes.

Adds an assertion in SLElement's matchesObject method to ensure that
matchesObject is never called on an SLElement instance with a nil
matching predicate.