Ensures that uiaPrefixes containing webviews are correct
Applies an accessibility identifier to any object within the
accessibilityChain that is located 2 indexes after a UIWebView. This is
done to ensure that the view located at this index - a UIWebBrowserView
is included in the UIAPrefix generated by Subliminal. This class
is one of several that UIAutomation includes in its accessibility
chain regardless of its label, name and value. For other classes  UIAutomation
treats this way we have created slAccessibilityCategories to ensure their
accessibilityName is not nil and they are included in the prefix.
We did not choose that solution in this case, because UIWebBrowserView
is a private class, and we do not wish to use private api's in
Subliminal.