Ensures UIAPrefix generates all required links
Adds a link in the generated uiaprefix for objects that are
uiaccessibilityelements, and have accessibility values, but do
not have accessibility names. This is necessary for matching
textfields in uiwebviews. Also assigns an accessibility identifier
to elements that are accessibility elements, but have neither
and accessibility name or an accesisbility value, in SLAccessibility's
adjustAccessibilityNamesInAccessibilityChain.