If test or test case set-up fails, the corresponding tear-down still executes.
This was already being done w/r/t test-case set-up, but the documentation
did not say so; both the documentation and the implementation have been
altered to guarantee that test tear-down is still executed after test set-up fails.

The changes in SLTest.m look big, but are for the most part just the existing
test case execution loop being wrapped in an if statement. See the beginning
and end of -run: for the real changes.