Tests and test cases may now be "focused". (fixes #66)
As an aid to debugging:
when a test is run, if any of its test cases are focused, only those test cases will run.

A test case is focused by prefixing its name with "focus_".
It is also possible to implicitly focus all test cases by prefixing their test's name with "Focus_".