Tests are now considered to support the current platform only if they have at least one test case which supports the current platform.
This ensures that tests without test cases, e.g. SLTest, will not run,
which will make for better logging when running +[SLTest allTests]. This
change also removes the implication that a test is an entity distinct
from its cases--it would have been meaningless for a test to have
supported the current platform while none of its test cases supported
the current platform.

Test cases have been added to the SLTests shared by the unit tests
as appropriate so that they will continue to run.