Refactors logic for creating UIAccessibility chains
Refactors logic used to determine whether or not an object
will exist in UIAutomation's accessibility chain into a couple
methods in NSObject (SLAccessibility). The existence of
slAccessibilityName is no longer used to determine whether or
not an object will appear in the accessibility chain, and the
accessibility identifier is no longer being set to a standard
replacement when an object must exist in the chain. slAccessibilityName
now has the sole responsibility of identifying elements to subliminal,
and accessibilityIdentifier is only ever set temporarily during
message sends to UIAutomation.

Subliminal's accessibility chains are being trimmed to match
UIAutomation's based on the methods shouldAppearInAccessibiltiyHierarchy,
shouldAppearInUIViewAccessibilityHierarchy, and
elementMockingSelfShouldAppearInAccessibilityHierarchy. These methods base their
return value on whether the object's parent is an accessibility element,
whether the object is an accessibility element, the existence of an
accessibilityIdentifier or certain accessibility traits,
and whether the object descends from one of several special case
classes. The logic in this method has been derived from trail and error
testing.