Added SLIntegrationTest, a base class for integration tests written against Subliminal itself.
Each SLIntegrationTest subclass is accompanied by a SLTestCaseViewController
subclass which provides the interface to be exercised by the test's cases.

In -setUpTest/-tearDownTest, SLIntegrationTest pushes/dismisses
an SLTestViewController which displays a list of the test cases it will execute.
(This is nothing more than an indicator of test activity, for the developer's sake.)

Then, in -setUpTestCaseWithSelector:/-dismissTestCaseWithSelector:,
SLIntegrationTest pushes/dismisses instances of its corresponding
SLTestCaseViewController class. SLTestCaseViewController provides a view
containing the elements to be exercised by the test's cases.