Ensures apphook targets are not retained
Nests the sending of apphook actions within an autoreleasepool.
This is necessary because when accessing the value property of the
SLWeakRef associated with the target, the value is retained and autoreleased.
Off of the main thread the base autorelease pool is not popped until the
end of the dispatch_block, which in this case is the end of testing. Nesting
in a @autoreleasepool{} ensures that the target is released immediately after
it is sent the message