Do not retain app hook targets during deregistration.
It is natural for targets to deregister themselves upon dealloc.
If the asynchronously-dispatched deregistration block captures
the target, it won't keep the target alive--as it's already set to die--
but the release of the captured reference at block close will cause a
segfault, as the reference will have by then become dangling.