Add integration test for Subliminal's dragWithStartPoint:endPoint: method
Adds an integration test to verify that Subliminal's dragWithStartPoint:endPoint:
method actually will cause a UIScrollView to drag.  These tests would
fail prior to the previous commit, where we used a drag duration of 0.2
in dragWithStartPoint:endPoint: because the scroll view's scrolling
velocity was so high at the end of the dragging gesture that the scroll
continued passed the expected content offset.

This test both verifies that dragWithStartPoint:endPoint: will perform
a slow drag rather than a fast swipe on a UIScrollView, and also that
the final content offset that results from the dragWithStartPoint:endPoint:
method is consistent with the distance between the input start and end
points.