While integration testing, uncaught exceptions are logged to Subliminal. (fixes #116)
So that those assertions are reported alongside other test failures.

We are able to log even exceptions thrown on the main thread
by spinning the run loop while waiting for UIAutomation.

Calls to -[SLTerminal eval:] are now serialized to prevent
a call to log an exception arising from the main thread (and then
dispatched to a global queue) from competing with a call to the terminal
from the test queue.