Refactored dragging test to clarify expectations.
1. We now drag the scrollview rather than the window,
for consistency, as we rely upon the scrollview to self-report the drag distance.
We use -[SLElement rect] to get the scrollview's rect.

2. The drag distance is described as such rather than as an "offset".

3. Investigation suggests that a tolerance is necessary when measuring the drag
distance not because UIAElement.dragInsideWithOptions is imprecise but because
the scrollview's delegate may not receive -scrollViewWillBeginDragging:
until dragging has occurred over a small distance (cf. the documentation).
So, we avoid suggesting that it is imprecise (pending further reports of errors)
by measuring the drag distance over a single run rather than iterating.

4. The visibility assertions have been factored out of the drag distance test into a
separate "simple" test which verifies just that the scrollview does scroll when you
drag it. The drag distance test is now called the "precise" test because it demonstrates
precisely what it means to drag in terms of the distance between the drag points.