-[SLElement waitUntil{Visible, inVisible}] account for changes in validity. (closes #122)
-waitUntilVisible: now allows for the element to be invalid at the beginning of the wait*,
so long as it is both valid and visible by the end of the wait.

-waitUntilInvisible: has been renamed to -waitUntilInvisibleOrInvalid: and, as the name
suggests, succeeds when the element becomes invisible _or_ invalid.

These changes clarify the intent of these methods: that the receiver should be,
in the case of the former, visible at the end of the wait, regardless of its start state;
and in the case of the latter, "not visible" at the end of the wait--which could mean
"invisible" or "invalid".

* It turns out that UIAElement.isVisible() reports NO, without throwing, if the element
is invalid--thus no changes to the implementation were strictly required to enable
this feature. However, we are safest--and echo Subliminal's "invalid elements are of
indeterminate visibility" semantics--if we explicitly check for isValid() before
isVisible().