-[SLElement isValid] no longer waits for the element to become valid. (fixes #126)
It should be able to be used to assess whether, at a discrete point in time,
an element is or is not valid. A client could easily use SLWaitUntilTrue and -isValid
together if they wanted (though it's more likely that they would use
-waitUntilVisible anyway).

Actions performed on the element (taps, etc.) still wait for the element
to become valid--but because we're not directly asking about validity;
we just want the action to succeed and are willing to give it a little leeway
toward that end.

UIAutomation appears to take this philosophy (-isValid doesn't use a timeout;
actions do) as well.