Require that all alerts be handled asynchronously, using handlers.
In-app purchase alerts cannot be handled synchronously. This is because
they are shown by a completely different process than the application under test.
Not only can they not be located by Subliminal like other elements--because
they are not part of the application's view hierarchy--but Subliminal cannot
even communicate with UIAutomation after they are shown, because UIAutomation
suspends execution when the application resigns active.

But, UIAutomation still calls the UIATarget.onAlert handler when such alerts are shown.
There, we have an opportunity to handle the alerts. This only requires that tests
register handlers before the alerts are shown.

Rather than provide two alert-handling mechanisms--a synchronous mechanism,
for regular alerts, and handlers, for XPC alerts--with according maintenance and
documentation costs, we remove the synchronous mechanism: SLAlerts are
no longer SLElements.