Log an error message if an alert is not handled and cannot be dismissed.
UIAutomation's default handler is supposed to throw an error if an alert cannot be
dismissed, but instead of just calls UIATarget.onAlert infinitely. We can't force the
tests to abort--if we throw an exception, Instruments will crash. If the tests are
being run via the command line, Instruments will eventually time out; if the tests
are being run via the GUI, the developer can abort.