-[SLElement waitUntilVisible:] and -waitUntilInvisibleOrInvalid: have become functions.
This mimics the syntax of the SLTest assertion macros, to feel more natural
--the tests are waiting for the element to change state, rather than asking the element
to wait for itself to change state--and to provide an opportunity to specify
a message to be logged if the assertion fails.

Note: In this commit, besides turning the methods into functions, I moved them down
to the bottom of SLElement.m (outside the @implementation block). For some reason,
Git has decided that I moved a bunch of stuff up instead.