Turn off auto-layout on all SLElementVisibilityTest XIBs.
For reasons unknown, auto-layout was squishing views to very short heights,
causing several of the visibility tests to fail. I couldn't figure out how
to edit the constraints to get it to stop doing that, but once I turned
off auto-layout, it was easy to remove the autoresizing masks it had
unnecessarily added.