SLElements only match objects that will appear in the accessibility hierarchy.
This prevents the search for a matching object from aborting prematurely,
e.g. at a accessibility container that shares a label with its child, but which
is not actually the accessible element, as in the included tests which match
web view child elements.

Note that -[NSObject willAppearInAccessibilityHierarchy] does not fully describe
the conditions under which objects such as intermediate containers will appear
in the accessibility hierarchy, i.e. the criteria by which SLAccessibilityPath
filters raw paths. However, it is difficult to replicate that criteria on a per-object
basis, as the raw paths are constructed, and it appears that -willAppearInAccessibilityHierarchy
*is* sufficient to evaluate paths' destinations.