Add visibility test for case where view is "covered" by a clear area
Adds a new test case to the SLElementVisibilityTest suite to test
the case where a view is visible but "covered" by another view,
where the covering view doesn't actually draw anything over the
center pixel of the covered view.

In this case UIAutomation will claim that the view is not visible,
but the view really *is* visible and this test verifies that our
own visibility tests get the expected result.