SLTerminal now supports the evaluation of scripts that evaluate to boolean or numeric values. (closes #7, #8)
We can only guarantee that strings and boolean/numeric values can be serialized
(as property-list values). We ensure that SLTerminal.js returns "null" for
results of other types.

Arrays and dictionaries (hashes) serialize fully if they contain only
property-list types, but if they don't, they serialize as empty collections. The
simplest guidance is to restrict scripts to returning strings, booleans, or numbers
(if anything). If at some point we need to support more complex message types we
should probably just switch to using JSON to communicate.