Do not attempt to warn if the accessibility inspector is enabled while unit testing.
As the Inspector should be irrelevant, and we can't just leave the existing code as-is
because the user directory path is different in unit tests than when the application
is running.

We use a run-time check for "is unit testing" so that we don't have to specially build
Subliminal when unit testing.