Determine some NSObject's visibility with UIAutomation
There are some NSObject subclasses that can be vended by
accessibility containers that are neither UIViews or UIAccessibiltiyElements,
such as WebAccessibilityObjects vended by UIWebBrowserViews.
Unfortunately, for these classes we do not have enough information about
how they are rendered to determine whether or not they are visible.

In the case of these objects we will need to use UIAutomation's
isVisible to determine if the objects are visible within their
containers. Since UIAutomation's isVisible implementation fails
in some cases to properly deal with UIView's we will also need
to check the visibility of the objects' accessibilityFrames on their
nearest UIView ancestors.