All SLLoggers must now be thread-safe.
`SLLogAsync` was originally intended as a utility for use by the application,
meaning that it would be the client's responsibility, if they used it,
to ensure that the shared logger they set was thread-safe.
However, Subliminal has since come to use `SLLogAsync` internally,
meaning that any shared logger must be thread-safe.