-[NSObject slReplacementAccessibilityIdentifier] no longer asserts that -slReplacementAccessibilityIdentifier has been loaded.
-slReplacementAccessibilityIdentifier must not be called unless
+loadSLReplacementAccessibilityIdentifier has been previously sent to the
receiver's class or a superclass--because if the replacement method is
called without having been loaded, and it tries to call -slTrueAccessibilityIdentifier,
that method won't yet be implemented--but the assertion will fail if a
never-bound class calls into a bound superclass' implementation
of -accessibilityIdentifier. (This can happen because SLStaticElements
don't bind their matching objects.)

Rather than try to provide an alternate guarantee and risk obfuscating
the code, I've added a bit to the documentation and comments. That should
be sufficient for debugging.