Subliminal distinguishes between "expected" and "unexpected" failures. (closes #5)
A failure is "unexpected" unless it was caused by an assertion failing.

This is consistent with SenTest, and both enables and encourages developers
to make specific claims in their tests, in the form of assertions.
By distinguishing between expected and unexpected failures,
they will be able to distinguish between those claims being false
and their tests having other issues.