Reorganized the project directory.
The file templates, along with `Integration Tests.xcconfig` and `SLTerminal.js`,
have been moved into a new `Supporting Files` directory (the latter item into
an `Instruments` subdirectory, the others into an `Xcode` subdirectory).
This leaves the `Documentation` directory entirely to the documentation
if it is generated.

The `Examples` directory has been renamed to simply `Example`, and the
example project moved to the top level within that directory. The example
project now uses a relative reference to the Subliminal project one level up,
rather than same-level symlinks to the project, one level up.