Add Rake tasks to build Subliminal's integration tests.
The test artifacts are archived to Supporting Files/CI/results,
which is now ignored by Git. It's a little weird to output
a build artifact into the Git repo but I don't know where else
to put it, and at least now it's all together with the rest of Subliminal.