The Rakefile now patches the trace template on installation to refer to SLTerminal.js at its location after installation.
With a valid reference to the file, Instruments will allow the script to be edited in-place,
although--because the developer would have to copy their edits out of the file--this is
done more to ensure that the template will work correctly (i.e. if it stopped caching
its script when it was saved) than to allow SLTerminal.js to be modified.