Implement better workarounds for instruments' authorization prompt.
aka http://openradar.appspot.com/radar?id=1544403.

Travis demonstrates that it is not sufficient to alter /`etc/authorization`.
Instead, we dismiss the prompt directly, using the user's login password.

We attempt to dismiss the prompt in two ways: first, if it appears as
an alert, we enter the password into, and dismiss, that alert;
second, if the alert does not appear (as in certain CI environments
like Travis), we enter the password at the command line.

Travis' configuration file has been modified to pass the current VM
password.