Assert `didHandleAlert` within a timeout.
Our previous assumption was that `didHandleAlert` could be checked
immediately after an alert was shown, because UIAutomation's `onAlert`
handler was called synchronously with the alert being shown; and JS is
single-threaded, so the script executed by `didHandleAlert` would block
on the execution of the `onAlert` handler.

However, for some reason, we were experiencing errors in CI
where a timeout was necessary--perhaps because in certain circumstances
`-[UIAlertView show]` does not immediately show the alert, or at least
cause the `onAlert` handler to be triggered.

Subliminal's guidance is now that `didHandleAlert` always be checked
within a timeout.