Move the meat of SLUIALogger to SLLogger and remove SLUIALogger (resolves #179)
Moves all the overrides froml SLUIALogger into SLLogger and removes
the separate SLUIALogger class.  Now the only logger class is
SLLogger, but SLLogger functions like SLUIALogger did before the
subclass was removed.

Changes references to SLUIALogger to SLLogger and removes comments
in SLLogger about the need for concrete subclases, moves SLLogger
out of the Logging group in the Subliminal project because it
would have been the only remaining class in that group, making
the group kind of silly.