Wait for a second before declaring the current test to be active
Causes SLTestViewController and SLTestsViewController to wait for
a full second after a view controller has come on screen before
assigning the variable that marks the pending test(case) as
the current one.  This way we can be more confident that the
test or test case is really ready to go before we start firing
off assertions.