Make dragging tests do nothing when TRAVIS environment variable is defined
These tests seem to work fine when we run them locally or in the
Jenkins CI environment, but they fail very frequently on Travis.
We have very little information to go on for the Travis failures
so for the time being we have to just disable these tests in that
particular environment.