`-[SLDevice setOrientation:]` delays before returning to ensure that UIDevice registers the new orientation.
Travis is exhibiting strange timing issues. It may be that in some circumstances,
`UIDevice` has not registered the new orientation by the time that
`UIATarget.setDeviceOrientation` returns.

Also ensure that the orientation will be set back to portrait even if a later rotation fails,
by moving the reset into `-tearDownTestCaseWithSelector:`.