Adds SLGeometry.
To be used for converting CGRects, CGPoints and CGSizes to their JavaScript
counterparts, and vice-versa. Updates existing code to use this.

The SLGeometry documentation won't actually work on Appledoc because it doesn't
support C functions -- but it's helpful when reading the header file, and hopefully
can be used with Clang's new documentation support.

Add SLGeometryUnitTests and SLGeometryTestViewController