Make SLTextField clear text before setting new text
When SLTextField takes the typeString path in its setText method,
it also needs to explicitly clear whatever text happens to be
in the text field already before starting typing.  Otherwise
the typeString call will just append the new text to the end
of whatever's currently in the text field (unless we luck out
with clearsOnBeginEditing).

This commit also modifies the testSetSearchBarText test case
to make sure that it exercises the scenario that previously
caused bad behavior.