Restore `-[SLTextField setText:]`'s use of `-[SLKeyboard typeString:]` in all cases.
This effectively reverts commit dc491fc72218a4d1c849f27a5482ec9adac620c9.
The UIAutomation bug described by that commit could not be reproduced:
`UIAKeyboard.typeString` appears to change the keyplane as necessary
to type special characters, across all device type and SDK combinations.

Pending further bug reports, using `-typeString:` is to be preferred
because it is simpler, because it most accurately simulates user interaction,
and because if there was a bug discouraging the use of `-typeString:`, a workaround
would have to be implemented not just in `-[SLTextField setText:]` but also
in `-setText:` as implemented by `SLWebTextField`, `SLTextView`, and `SLWebTextView`.

A test case has been added to verify that `-[SLKeyboard typeString:]` changes
the keyplane as necessary to type special characters.