Restart instruments if it fails to launch the target.
This will hopefully resolve failures that we see somewhat more than occasionally
on Travis such as https://travis-ci.org/inkling/Subliminal/builds/9926376#L1189,
where--within the same run--instruments may fail to launch the app on one platform
and then run the others successfully. Since this seems to be a momentary failure,
instruments will hopefully succeed if we retry, after a short delay for good measure.
We've not been able to reproduce this error locally or in other CI environments,
unfortunately. If this fix does not successfully work around the problem we will try
something else.