`-[SLElement logElement]` now uses `-slAccessibilityDescription` to log the underlying object. (fixes #65)
This ensures that the user will be able to see the element's unbound name
(because the accessibility description can be calculated without having to talk
to UIAutomation), and actually allows us to log additional information beyond
that provided by UIAutomation.

However, `-logElementTree` still uses UIAutomation because UIAutomation
will log just those elements in the accessibility hierarchy, which would be expensive
for us to determine.