Reset the simulator and rebuild the app before restarting instruments.
Delaying for a longish period of time before restarting (f7a0ff565eab437d04d145aabf44798354422fdd)
did not work: https://travis-ci.org/inkling/Subliminal/builds/10263900#L2210.
Perhaps `instruments` is rejecting the build for some reason.
Certainly, when we have seen a similar error locally, it is because the build
was invalid (for instance, it was built for the device but attempted to be run
in the simulator).