`SLTest` and `SLTestController` use `SLLogger` to log exceptions occurring during tests. (refs #72)
This entails moving `SLTest`'s exception logging routine to `SLLogger`,
while the determination of whether the exception was expected,
and whether call site information should be attached, remains in `SLTest`.

This change is made to unify `SLTestController`'s exception logging with
that of `SLTest`.

Because `SLTest` and `SLTestController` no longer log exception error messages
directly--they call `-[SLLogger logException:expected:]` rather than `-logError:`--
in order to continue to test that those error messages include call site information,
the `SLTestTests`' logger mock must become a partial mock (so that the actual
logger will call `-logError:`). It was actually unnecessary to stub the logger
for the purposes of "preventing the framework from trying to talk to UIAutomation,"
as the comment said--that is accomplished by stubbing the terminal below that.