If a test fails in both test set-up and tear-down, both exceptions will be logged.
We had previously ignored the second exception on the basis that it was superfluous
--expected somehow on the basis that set-up had already failed. But,
if a test fails in both test *case* set-up and tear-down, we log both those exceptions;
and it's probably better to err on the side of more information.

Since there is potentially both an exception from test set-up _and_ an exception
from test tear-down to log, `-[SLTest runAndReportNumExecuted:failed:failedUnexpectedly:]`
can no longer throw such exceptions for the test controller to log, but rather logs those
exceptions itself and returns `YES` or `NO` to indicate whether the test finished or aborted.
This actually makes the test execution structure a fair bit simpler.

Because `SLTest` is now logging all test exceptions itself,
it no longer needs to expose `+exceptionWasExpected:`.