Instances of `SLElement` can be directed to double-check their validity using UIAutomation.
To diagnose bugs in `SLAccessibilityPath` and/or `NSObject (SLAccessibilityHierarchy)`,
where Subliminal believes an element to be valid but incorrectly identifies it to UIAutomation,
as it is not guaranteed that UIAutomation will raise an exception if an accessibility path
is invalid.

This setting is disabled by default because it is more likely that there is a bug in
a particular test than in UIAutomation and because enabling it will negatively
affect the performance of the tests (by requiring an extra call to UIAutomation).