Subliminal now reloads the accessibility hierarchy as necessary while matching.
In certain circumstances, accessibility elements may become "stale", in that they match
views which are no longer in the hierarchy. When these elements are queried for their
accessibility information, their container will reload _all_ its child elements.
This could result in Subliminal finding a path to an element only to have the process
of finding that path invalidate the path! We now preemptively reload the accessibility
hierarchy while retrieving accessibility elements to ensure that accessibility paths remain
valid.