`SLElement` methods that dynamically identify an element within a given timeout retry identification if a valid element becomes invalid as long as time remains.
Rather than assuming that a valid element will remain valid while it is being used.
During animations, when the view hierarchy is undergoing rapid modification,
it's possible for Subliminal to identify a valid and tappable element, only for that
element to have been replaced in the accessibility hierarchy by the time that
UIAutomation goes to manipulate the element.