`set_simulator_device` only quotes the device type if necessary.
Both the Xcode 4 and 5 Simulators reject e.g. "'iPhone'" . It went unnoticed
that `set_simulator_device` wasn't working for "iPhone" and "iPad" because
`subliminal_test` forces the project to build using particular device families
and "iPhone" and "iPad" are the default device types for their respective families.

Investigation reveals that it's not whitespace but rather punctuation in plist values
that requires quoting so that's updated too--now `set_simulator_device` will
be able to set the Xcode 5 Simulator to "iPad Retina".