Programmatically scroll scroll views in the iOS 7 Simulator.
In the iOS 7 Simulator, scroll views' pan gesture recognizers fail to cause those views
to scroll in response to UIAutomation drag gestures: a scroll view (and its pan gesture
recognizer) receives the touches involved in the gesture, but something inscrutable
goes wrong in the view's response to that gesture.

Subliminal uses a private UIAccessibility API to programmatically scroll the views
(with a higher degree of fidelity than `-setContentOffset:` would afford, e.g. the
private API notifies the delegate of appropriate scroll events as if the view
was being dragged). This method's usage need not be obfuscated because
it is compiled for the Simulator alone.