Create a test matrix. (fixes #96)
Every row in `env:matrix:` specifies a build.

Configuration documentation is here:
http://about.travis-ci.org/docs/user/build-configuration/#The-Build-Matrix.
Unfortunately, it's not currently possible to parameterize _among_ environment variables,
only between environment variables and, say, runtimes; so we must explicitly
specify the options to be used for every build vs. having Travis construct
a matrix from possible values of `TEST_COMMAND` and `TEST_SDK`.

We are able to use `env:global:` to attach the login password to every row.