Simplify the test of the uniformity of Subliminal's test distributions.
After another failure of this test on Travis, I took another look at the chi-square code
and I don't think it's ever been correct and/or we've ever been applying it correctly.
We've been claiming that there is at least a 5% chance (recently changed to 7.5%)
that the distribution is uniform, but a chi-square test, when applied to a uniform distribution,
usually supports the claim that there is *less than* a 5% chance that the distribution
*is not* uniform. That is, we've been making a very weak claim. And it turns out
that the significance values found by the chi-square code have been all over the place:
often above 7.5%, but sometimes below (which is when this fails).

I'm still fairly certain that our randomization code is correct, i.e. our distributions
are uniform, and that we should be able to apply a chi-square test to support
that claim. But I don't think that it's worth trying to patch up this code.
I've replaced the chi-square test with a simpler claim which ought to verify
the practical use and validity of our randomization.