Document issue about swiping table view cell [skip ci]
In prior versions of iOS, UITableViewCell instances used a discrete
swipe gesture recognizer to implement the ‘swipe-to-delete’ behavior
and show the deletion confirmation button.

With iOS 7 the internal view hierarchy of a cell has been modified:
a scroll view has been inserted between the UITableViewCell and the
content view. It is responsible for tracking the swiping gesture and
set the inset for the whole content view.

Since Subliminal cannot drag scroll views in the iOS 7 Simulator for
the time being, it cannot swipe a table view cell either. Let’s add
this as a known issue in README.md.