Subliminal logs uncaught exceptions using a dedicated log method.
So that uncaught exception logging may be tested directly.

The current test and test case are no longer logged as part of the
uncaught exception message. Saying that the exception occurred
"during" the test might imply that the exception occurred as a result
of the test (rather than in the application while the test was ongoing);
such information can be deduced from the logs prior to the exception;
and this reduces the state for the test controller and tests to maintain.