Get the path for the deepest matching element
We usually want to grab the path for the deepest (leaf) matching object
in the tree; this is especially true in iOS 7 since Apple changed
the accessibility hierarchy and inserted intermediate views or
accessibility elements in between.

Let’s modify `-rawAccessibilityPathToElement:favoringSubviews:` to
search for the deepest matching object instead of return the first the
method encounters.